requirements.txt
Django>=4.2,<5.0
djangorestframework>=3.15
djangorestframework-simplejwt>=5.3
drf-spectacular>=0.27
django-filter>=24
python-decouple>=3.8
pytest-django>=4.8
pytest>=8.0
